% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angleList.R
\docType{data}
\name{angleList}
\alias{angleList}
\title{Example output from TurningAngles()}
\format{
list
}
\usage{
angleList
}
\description{
Example output from the \code{\link{turningAngles}} function. Angles calculated using
the \code{\link{tracks}} dataset and turningAngles() with max_hr=24. Note that the
turningAngles() default is normally max_hr=240; however, this creates an unnecessarily
large file for an example dataset so max_hr=24 was used instead.
}
\keyword{datasets}
