% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angleListAll.R
\docType{data}
\name{angleListAll}
\alias{angleListAll}
\title{Example output from turningAngles()}
\format{
list
}
\usage{
angleListAll
}
\description{
Example output from the \code{\link{turningAngles}} function. Angles calculated using
the \code{\link{tracks}} dataset and turningAngles() default parameters.
}
\keyword{datasets}
