% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDisp.R
\name{calcDisp}
\alias{calcDisp}
\title{Calculate displacements}
\usage{
calcDisp(
  species_df,
  min_hr = 24,
  max_hr = 240,
  interval_hr = 24,
  range_hr = 6,
  verbose = FALSE
)
}
\arguments{
\item{species_df}{A data frame containing location data in rows. Columns have the following headers: "ref", "lon", "lat", "day".
"ref" is the unique id number for each individual (e.g., their satellite tag number formatted as an integer),
"lon" and "lat" are the longitude and latitude of each position estimate in decimal degrees in numeric format,
"day" is the datetime stamp for each location estimate in POSIXct format following yyyy-mm-dd hh:mm:ss.
See attached sample data \code{\link{tracks}}.}

\item{min_hr}{Minimum number of hours to consider for calculations. Default is 24 hours.}

\item{max_hr}{Maximum number of hours to consider for calculations. Default is 240 hours.}

\item{interval_hr}{Time interval (in hours) used to set intervals between min_hr and max_hr. Default is 24 hours.}

\item{range_hr}{Range (in hours) converts interval_hr into a time window (interval_hr +/-  range_hr) so the
code can identify location estimates that are close to, but not exactly separated by the interval_hr input value.
If multiple location estimates fall within this time window the location estimate closest to the interval_hr input value
will be used for calculations. For example, if interval_hr = 24 and range = 6, the algorithm will search for
locations spaced 18 to 30 hours apart. Default is 6.}

\item{verbose}{Logical. If TRUE, informative messages describing the number of displacements identified for 
each time window are displayed. Default is FALSE.}
}
\value{
A list containing the displacements in km recorded for each time window. Each list element corresponds with the time
windows set (i.e., the first list element is the first time window).
}
\description{
This function allows you to calculate the displacement distances travelled by individuals over set time windows.
}
\examples{
\donttest{

calcDisp(tracks, min_hr=24, max_hr=240, interval_hr=24, range_hr=6, verbose=TRUE)

}
}
