% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateDisplacements.R
\name{CalculateDisplacements}
\alias{CalculateDisplacements}
\title{Calculate Displacements}
\usage{
CalculateDisplacements(
  species_df,
  min_hr = 24,
  max_hr = 240,
  interval_hr = 24,
  range_hr = 6
)
}
\arguments{
\item{species_df}{A data frame containing location data (rows) and columns with the following headers: "ref", "lon", "lat", "day". "ref" is the unique id number for each animal
(e.g., their satellite tag number; integer format). "lon" and "lat" are the longitude and latitide of each position estimate in decimal degrees (numeric format). "day"
is the datetime stamp for each location estimate (POSIXct format following yyyy-mm-dd hh:mm:ss). See XXX data frame for an example.}

\item{min_hr}{Minimum number of hours to consider for displacement calculations. Default is 24 hours.}

\item{max_hr}{Maximum number of hours to consider for displacement calculations (default is 240 hours)}

\item{interval_hr}{Time interval (in hours) used to identify time period intervals between min_hr and max_hr (default is 24 hours)}

\item{range_hr}{Range (in hours) applied to interval_hr. This value helps the algorithm identify location estimates that are close to, but not exactly separated by the interval_hr. If multiple location estimates fall within this range the location estimate closest to the interval_hr will be used for calculations.}
}
\value{
A list containing the displacements (distances traveled in km) recorded for each time period
}
\description{
This function allows you to calculate the displacement distances traveled by individuals over set time windows.
}
\examples{
CalculateDisplacements(species_df)
CalculateDisplacements(species_df, min_hr=24, max_hr=240, interval_hr=24, range_hr=6)
}
