% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommunityMap.R
\name{CommunityMap}
\alias{CommunityMap}
\title{Map Infomap Communities}
\usage{
CommunityMap(infomap_object, core_communities)
}
\arguments{
\item{infomap_object}{output from the \code{\link{InfomapCommunities}} function}

\item{core_communities}{Vector of communities to be plotted in map. Default is NULL}
}
\value{
A map illustrating Infomap communities
}
\description{
This function allows you to create a map of infomap communities calculated using the \code{\link{InfomapCommunities}} function
}
\examples{
CommunityMap(infomap_object)
CommunityMap(infomap_object, core_communities)
}
