% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compDist.R
\name{compDist}
\alias{compDist}
\title{Identify the best-fit distribution for data}
\usage{
compDist(input, distResults, force_AICc = FALSE)
}
\arguments{
\item{input}{List of values used to evaluate and compare distribution fits generated by fitDist()}

\item{distResults}{List output from the \code{\link{fitDist}} function containing a dataframe of fit results (element 1) and a normalisation record (element 2)}

\item{force_AICc}{Force function to calculate AICc scores instead of AIC scores when n/K is > 40. Default is FALSE.}
}
\value{
A data frame with that contains the summary statistics for each distribution fit (from the \code{\link{fitDist}} function) as well as
the AICc/AIC scores and weighted AICc/AIC scores (wAICc/wAIC) for each distribution fit.
}
\description{
This function allows you to determine if a power law, exponential, or log-normal distribution best-fit a probability density function
of the input data using weighted Akaike Information Criterion (AIC). These fits use input data in conjunction with dmin and parameter values that were
previously calculated with the \code{\link{fitDist}} function. By default, this function will calculate
AICc scores (AIC scores corrected for small sample sizes) if n/K is <= 40 for the largest value of K, where n = sample size (nTail) and
K = number of parameters in the model (see Burnham and Anderson (2004) for further details, <doi:10.1177/0049124104268644>). However,
if force_AICc = TRUE AICc scores will be calculated regardless of n/K.
}
\examples{
compDist(disp, distResultsAll, force_AICc=FALSE)
}
