% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displacements.R
\docType{data}
\name{displacements}
\alias{displacements}
\title{Sample displacement data to demonstrate PhysMove functions}
\format{
A list with 10 elements
}
\usage{
displacements
}
\description{
A list containing the displacements in km recorded for each time window using CalcDisp default parameters
}
\keyword{datasets}
