% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DispPDFplot.R
\name{DispPDFPlot}
\alias{DispPDFPlot}
\title{Create probability density function plots (PDFs) of displacements}
\usage{
DispPDFPlot(displacements, colours = rainbow, legend = c(TRUE, "bottomleft"))
}
\arguments{
\item{displacements}{The output from the \code{\link{CalcDisp}} function.}

\item{colours}{Colour(s) for points in pdf plot. Valid options include: base R (grDevices) color pallets (e.g., colours=rainbow),
specific colours (e.g., colours=c("Navy"), or colours=c("red","blue")), or hex numbers (e.g., colours=c("#FF0000")). Default is rainbow.}

\item{legend}{Add a legend to the plot when TRUE and change the position of the legend. Default is legend=c(TRUE, "bottomleft")}
}
\value{
Probability density function (PDF) plots of raw and normalized displacements.
}
\description{
This function allows you to plot probability density functions (PDFs) of raw displacements and
displacements normalized by the mean displacement for each time window. Displacements must have been
previously calculated with the \code{\link{CalcDist}} function.
}
\examples{
DispPDFPlot(displacements)
DispPDFPlot(displacements, colours=rainbow, legend=c(TRUE, "bottomleft"))
}
