% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distResultsExp.R
\docType{data}
\name{distResultsExp}
\alias{distResultsExp}
\title{Example output from FitDist() using the best-fit d min for an exponential distribution}
\format{
list
}
\usage{
distResultsExp
}
\description{
Example output from the \code{\link{FitDist}} function. Distribution fits calculated using the \code{\link{tracks}} dataset
and FitDist() with set_dmin=1.649160, which is the best-fit d min for an exponential distribution for the \code{\link{tracks}} dataset.
First list element contains a dataframe of distribution results (d min, parameters, etc) and the second list element is a record
of if the data were normalized or not, which is needed for \code{\link{CompDist}} and \code{\link{PlotDist}}).
}
\keyword{datasets}
