% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\title{Entropy of trajectories}
\usage{
entropy(species_df, gridCell = 0.25, histPlot = TRUE)
}
\arguments{
\item{species_df}{A data frame containing location data in rows. Columns have the following headers: "ref", "lon", "lat", "day".
"ref" is the unique id number for each animal (e.g., their satellite tag number formatted as an integer),
"lon" and "lat" are the longitude and latitude of each position estimate in decimal degrees in numeric format,
"day" is the datetime stamp for each location estimate in POSIXct format following '%Y-%m-%d %H:%M:%S'.
See attached sample data \code{\link{tracks}}.}

\item{gridCell}{Grid cell size in degrees. Default is 0.25.}

\item{histPlot}{Plot a histogram of the normalised entropy values. Default is TRUE.}
}
\value{
Data frame of the normalised entropy values for each trajectory as well as the individual entropy
values (not normalised) and the number of cells each trajectory visited. If histPlot=TRUE a histogram of the normalised entropy scores is created.
}
\description{
This function calculates the normalised entropy of individual trajectories based on the probability distribution of location observations across grid cells.
Normalised entropy scores are calculated by dividing individual entropy scores by the log number of cells each trajectory visited,
providing insight to how ordered or disordered the trajectories were.
Values close to 1 indicate high entropy (disordered trajectories), while values closer to 0 indicate low entropy (ordered trajectories).
A pdf plot of the normalized entropy values can be created with the \code{\link{plotPDF}} function.
}
\examples{
\donttest{

entropy(tracks, gridCell=0.25, histPlot=TRUE)

}
}
