% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropyResults.R
\docType{data}
\name{entropyResults}
\alias{entropyResults}
\title{Example output from Entropy}
\format{
dataframe
}
\usage{
entropyResults
}
\description{
Output from \code{\link{Entropy}} function using default parameters and the \code{\link{tracks}} dataset
}
\keyword{datasets}
