% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfomapCommunities.R
\name{InfomapCommunities}
\alias{InfomapCommunities}
\title{Create Link List for Infomap Analysis}
\usage{
InfomapCommunities(
  species_df,
  gridcell = 0.25,
  hours = 24,
  range_hr = 6,
  infomap = TRUE,
  tpm = FALSE
)
}
\arguments{
\item{species_df}{A data frame containing location data (rows) and columns with the following headers: "ref", "lon", "lat", "day". "ref" is the unique id number for each animal
(e.g., their satellite tag number; integer format). "lon" and "lat" are the longitude and latitide of each position estimate in decimal degrees (numeric format). "day"
is the datetime stamp for each location estimate (POSIXct format following yyyy-mm-dd hh:mm:ss). See XXX data frame for an example.}

\item{gridcell}{Grid cell size in degrees. Default is 0.25.}

\item{hours}{Number of hours to consider movements for. Default is 24.}

\item{range_hr}{Range (in hours) applied to hours This value helps the algorithm identify location estimates that are close to, but not exactly separated by the interval_hr. If multiple location estimates fall within this range_hr the location estimate closest to the interval_hr will be used for calculations.}
}
\value{
A matrix containing the link list needed to calculate Infomap communities using https://www.mapequation.org/infomap/#Install, and a data frame of the LinkListCellNumbers that is used with the \code{\link{InfomapCoords}} function to interpret the output from Infomap.
}
\description{
This function allows you to create a link list for the Network community detection Infomap (https://www.mapequation.org/infomap/#Install)
When function is finished please export LinkList as .txt file following "write.table(LinkListOutput, "LinkList2.txt", sep="/t", row.names = FALSE, col.names = FALSE)" and upload .txt file at https://www.mapequation.org/infomap/#Install
Important settings to consider when configuring analysis for Infomap are: '-i link-list -k --tree -d'
'-i link-list' identifies the input at a link list, '-k' enables self-links, '--tree' ensures the output is in .tree format, and '-d' assumed directed links."
}
\examples{
InfomapCommunities(species_df)
InfomapCommunities(species_df, gridcell=0.25, hours=24, range_hr=6, infomap=TRUE, tpm=FALSE)
}
