% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfomapLatLong.R
\name{InfomapCoords}
\alias{InfomapCoords}
\title{Convert Infomap Nodes to Coordinates
Assigns coordinates to the formatted output from Infomap using the LinkListCellNumber data frame and grid outputs from the \code{\link{LinkList}} function.
To format the Infomap output:
1) Download .tree output from Infomap website (https://www.mapequation.org/infomap/#Install):
1) Open file in excel
2) Delete summary text before results (cell A1 should contain path data in #:#:# format)
3) Using "text to columns" tool, set "space" as delimiter, and format column with path information as "text" (default is "General")
4) Save as "Network.csv"
5) Import to R as data.frame and add headers "path", "flow", "name", and "node" to columns 1 to 4, respectively}
\usage{
InfomapCoords(Network)
}
\arguments{
\item{Network}{Formatted output from https://www.mapequation.org/infomap/#Install.}
}
\value{
InfomapCoords dataframe containing
}
\description{
Convert Infomap Nodes to Coordinates
Assigns coordinates to the formatted output from Infomap using the LinkListCellNumber data frame and grid outputs from the \code{\link{LinkList}} function.
To format the Infomap output:
1) Download .tree output from Infomap website (https://www.mapequation.org/infomap/#Install):
1) Open file in excel
2) Delete summary text before results (cell A1 should contain path data in #:#:# format)
3) Using "text to columns" tool, set "space" as delimiter, and format column with path information as "text" (default is "General")
4) Save as "Network.csv"
5) Import to R as data.frame and add headers "path", "flow", "name", and "node" to columns 1 to 4, respectively
}
\examples{
InfomapCoords(Network)
}
