% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infomapResult.R
\docType{data}
\name{infomapResult}
\alias{infomapResult}
\title{Example output from InfomapCommunities()}
\format{
list
}
\usage{
infomapResult
}
\description{
Example output from the \code{\link{InfomapCommunities}} function. Results calculated using the
\code{\link{tracks}} dataset and InfomapCommunities() default parameters. The first list element
contains the Infomap results (i.e., an Infomap monolayer object) and the second list element includes a
transition probability matrix (tpm).
}
\keyword{datasets}
