% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Occupancy_20210317.R
\name{Occupancy}
\alias{Occupancy}
\title{Occupancy}
\usage{
Occupancy(species_df, gridcell = 0.25, map = TRUE, plot = FALSE, nb = 20)
}
\arguments{
\item{species_df}{A data frame containing location data (rows) and columns with the following headers: "ref", "lon", "lat", "day". "ref" is the unique id number for each animal
(e.g., their satellite tag number; integer format). "lon" and "lat" are the longitude and latitide of each position estimate in decimal degrees (numeric format). "day"
is the datetime stamp for each location estimate (POSIXct format following yyyy-mm-dd hh:mm:ss). See XXX data frame for an example.}

\item{gridcell}{Grid cell size in degrees. Default is 0.25.}

\item{map}{Create a map illustrating where occupancy occurs. Default is TRUE.}

\item{plot}{Create line plot of the probability density function of occupancy Default is TRUE.}

\item{nb}{Number of bins, used to determine the size of the bins used to calculate occupancy for plot. Default is 20.}
}
\value{
"MyOcc" data frame containing the location and corresponding occupancy data used in the plots, as well as a line plot and/or a map, if desired.
}
\description{
This function allows you to calculate the occupancy patterns of satellite tagged animals. and plot a pdf of the occupancy scores
}
\examples{
Occupancy(species_df)
Occupancy(species_df,grid=4,plot=TRUE,nb=20,map=TRUE)
}
