% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupancyDF.R
\docType{data}
\name{occupancyDF}
\alias{occupancyDF}
\title{Example output from Occupancy function for pdfPlot function}
\format{
data frame
}
\usage{
occupancyDF
}
\description{
Re-formatted results from Occupancy function using default parameters. Occupancy output is normally in list format but it was converted to a data frame for
simplicity with pdfPlot example (i.e., occupancy.df <- as.data.frame(Occ[[1]]), where Occ was the output from the Occupancy function.
}
\keyword{datasets}
