% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupancyResults.R
\docType{data}
\name{occupancyResults}
\alias{occupancyResults}
\title{Example output from Occupancy function for pdfPlot function}
\format{
data frame
}
\usage{
occupancyResults
}
\description{
Results from Occupancy function using Occupancy() default parameters.
}
\keyword{datasets}
