% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotAngles.R
\name{PlotAngles}
\alias{PlotAngles}
\title{Plot turning angles with a circle plot}
\usage{
PlotAngles(angleList, timePlot = "all", colours = rainbow, legend = TRUE)
}
\arguments{
\item{angleList}{List of angles calculates with the \code{\link{TurningAngles}} function.}

\item{timePlot}{Plot angles from all time windows or only plot angles from one specific time window. For example,
timePlot=1 will only plot angles from the first time window while timePlot="all" will plot all time windows.
Default is timePlot="all".}

\item{colours}{Colour(s) for lines in circle plot Valid input options include: base R (grDevices) color pallets (e.g., colours=rainbow),
RColorBrewer palettes (e.g., colours="Dark2"), and colour names or hex numbers (e.g.,colours=c("darkred", "#4682B4", "#00008B", "darkgreen")). Note that
grDevices color pallets are functions and do not use quotations. If the palette does not have enough distinct colours to match the lines being plotted the function will
automatically create a continuous pallet with the colours provided. Default is rainbow.}

\item{legend}{Add a legend to the circle plot. Default is TRUE.}
}
\value{
Circle plot of the angles calculated with the \code{\link{TurningAngles}} function and the data used to create the circle plot.
}
\description{
This function allows you to create a circle plot illustrating the frequency of turning angles from
the \code{\link{TurningAngles}} function.
}
\examples{
PlotAngles(angleList)
PlotAngles(angleList, timePlot="all", colours=rainbow, legend=TRUE)
}
