% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDispPDF.R
\name{plotDispPDF}
\alias{plotDispPDF}
\title{Create probability density function (PDF) plots of displacements}
\usage{
plotDispPDF(displacements, normalise = TRUE, colours = rainbow, legend = TRUE)
}
\arguments{
\item{displacements}{Displacements in list format (e.g., the output from \code{\link{calcDisp}}).}

\item{normalise}{Normalise the displacements by the mean displacement for each time window. Default is TRUE.}

\item{colours}{Colour(s) for plot points. Valid input options include: base R (grDevices) colour pallets (e.g., colours=rainbow), RColorBrewer
palettes (e.g., colours="Dark2"), and colour names or hex numbers (e.g.,colours=c("darkred", "#4682B4", "#00008B", "darkgreen")). Note that grDevices colour
pallets do not use quotations. If the palette does not have enough distinct colours to match the communities being plotted the function will automatically
create a continuous pallet with the colours provided. Default is rainbow.}

\item{legend}{Add legend with legend=TRUE. Default is TRUE.}
}
\value{
Produces probability density function (pdf) plots of displacements. Invisibly returns a data frame containing the calculated pdf values,
corresponding displacements, and time window identifiers used to generate the plot. Assign the output to an object to access these data.
}
\description{
This function allows you to plot probability density functions (pdfs) of displacements. Displacements must be in
list format where each list element corresponds to displacements calculated over a specific time window, which is the default output
format from the \code{\link{calcDisp}} function.
}
\examples{
plotDispPDF(disp)
}
