% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDistributions_20210319.R
\name{PlotDist}
\alias{PlotDist}
\title{Plot best-fit distributions to CCDF of displacements}
\usage{
PlotDist(
  Displacements,
  lines = TRUE,
  colours = c("#D55E00", "#DDCC77", "#0072B2"),
  distPlot = FALSE,
  legend = c(TRUE, "bottomleft")
)
}
\arguments{
\item{Displacements}{Input the output from \code{\link{CalculateDisplacements}} function.}

\item{lines}{Add fit lines based on parameters calculated with \code{\link{FitDist}} function. Default is TRUE.}

\item{colours}{Colours for each fit line. Default includes: red (#D55E00), yellow (#DDCC77), and blue (#0072B2).}

\item{legend}{Adds legend to plot and specifies legend location. Default is c(TRUE, "bottomleft").}
}
\value{
Complementary cumulative distribution function (CCDF) plot with fit lines (if  desired).
}
\description{
This function allows you to create a complementary cumulative distribution function (CCDF) plot with fit lines (if  desired) for the distribution fits calculated with the  \code{\link{FitDist}} function.
normalizes all of the displacements by dividing each displacement by the mean displacement from it's corresponding time period and combines the results into a vector of Normalized values.
}
\examples{
PlotDist(Displacements, Lines=TRUE, colours=c("#D55E00", "#DDCC77", "#0072B2"))
PlotDist(Displacements)
}
