% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRandomTracks.R
\name{plotRandomTracks}
\alias{plotRandomTracks}
\title{Plot Randomised Tracks}
\usage{
plotRandomTracks(
  species_df,
  ref = NULL,
  randomResults,
  numPlot = 1:5,
  colours = c("black", "grey70"),
  tracks = TRUE,
  startCol = "red",
  endCol = "blue",
  legend = TRUE
)
}
\arguments{
\item{species_df}{A data frame containing location data in rows. Columns have the following headers: "ref", "lon", "lat", "day".
"ref" is the unique id number for each animal (e.g., their satellite tag number formatted as an integer),
"lon" and "lat" are the longitude and latitude of each position estimate in decimal degrees in numeric format,
"day" is the datetime stamp for each location estimate in POSIXct format following '%Y-%m-%d %H:%M:%S'.
See attached sample data \code{\link{tracks}}.}

\item{ref}{Reference number of track from species_df to plot.}

\item{randomResults}{Result from \code{\link{randomise}} function. Note that the attached example dataset \code{\link{randomResults}}
was generated with randTrack=1 (to keep the example data small), so it only contains 1 randomised track; set numPlot=1 when using it.}

\item{numPlot}{Number of randomised tracks to plot. The randomised tracks were consecutively numbered from 1 to however many you set in the
\code{\link{randomise}} function. The input value can either be any of these individual numbers (e.g., 23), or a range of numbers (e.g., 1:10),
which will plot all of the random tracks created within the range. Default is 1:5. Must not exceed the number of randomised tracks
available in randomResults (e.g., numPlot=1 if randomResults was generated with randTrack=1).}

\item{colours}{Colours to plot points from original and randomised tracks, respectively. Default is colours=c("black","grey70").}

\item{tracks}{Add track lines to the plot. Default is TRUE.}

\item{startCol}{Colour for origin location. startCol=NULL will cause the symbology of the origin location to match the symbology of the rest
of the original track. Default is startCol="red".}

\item{endCol}{Colour for destination location. endCol=NULL will cause the symbology of the destination location to match the symbology of the rest
of the original track. Default is endCol="blue".}

\item{legend}{Add legend with legend=TRUE (default).}
}
\value{
Produces a plot showing the original and randomised track locations. Invisibly returns a data frame containing the randomised track
data used to create the plot. Assign the output to an object to access these data.
}
\description{
Plot locations from original and reshuffled tracks using RandomisedLat and RandomisedLong outputs from the \code{\link{randomise}} function
}
\examples{
plotRandomTracks(tracks, ref=1, randomResults=randomResults, numPlot=1)
}
