% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predictability_20210319.R
\name{Predictability}
\alias{Predictability}
\title{Predictability}
\usage{
Predictability(
  species_df,
  start_val = 0.99,
  hist = TRUE,
  pdfplot = FALSE,
  nb = 40,
  legend = c(TRUE, "topleft")
)
}
\arguments{
\item{species_df}{A data frame containing location data (rows) and columns with the following headers: "ref", "lon", "lat", "day". "ref" is the unique id number for each animal
(e.g., their satellite tag number; integer format). "lon" and "lat" are the longitude and latitide of each position estimate in decimal degrees (numeric format). "day"
is the datetime stamp for each location estimate (POSIXct format following yyyy-mm-dd hh:mm:ss). See XXX data frame for an example.}

\item{start_val}{Starting value used to find a root for the Limit of Predictability equation. If no value is provided function will loop through values starting at 0.99 and decreasing by 0.01 at each iteration until acceptable root values are identified. Default is 0.99}

\item{nb}{Number of bins, used to determine the size of the bins used to calculate predictability frequencies for plots. Default is 40.}

\item{plot}{Line plot illustrating the predictability of each trajectory.  Default is TRUE.}
}
\value{
Predictability vector containing predictability values for each trajectory, line plot (if desired).
}
\description{
This function allows you to calculate the predictability of each trajectory based on each individual's entropy. This function requires "Ind_Entropy","Occurrences", and "NormalizedEntropy" from \code{\link{Entropy}} function
}
\examples{
Predictability(species_df)
Predictability(species_df, start_val=0.99, plot=TRUE, nb=40)
}
