% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reshuffle.R
\name{Randomize}
\alias{Randomize}
\title{Randomizing trajectories}
\usage{
Randomize(
  species_df,
  RandTraj = 500,
  gridcell = 0.25,
  plot = TRUE,
  pchtype = 19,
  colours = c("black", "black"),
  title = "",
  legend = c(TRUE, "topleft")
)
}
\arguments{
\item{species_df}{A data frame containing location data (rows) and columns with the following headers: "ref", "lon", "lat", "day". "ref" is the unique id number for each animal
(e.g., their satellite tag number; integer format). "lon" and "lat" are the longitude and latitide of each position estimate in decimal degrees (numeric format). "day"
is the datetime stamp for each location estimate (POSIXct format following yyyy-mm-dd hh:mm:ss). See XXX data frame for an example.}

\item{RandTraj}{Number of randomised trajectories per individual}

\item{gridcell}{grid cell size in degrees. Default is 0.25.}

\item{Colour}{Colour of line for data. Default="Navy".}
}
\value{
RandomizedResults data frame that includes the number of cells visited by each individual's original and reshuffled trajectories,RandomizedLat and RandomizedLon matricies for plotting tracks (see \code{\link{PlotRandomTracks}} function for further details),
and a plot illustrating the number of cells visited by the original and reshuffled trajectories, if desired.
}
\description{
Randomizing trajectories to break any existing spatial or temporal correlations between location estimates. Here, the sequences of the displacements for each observed trajectory are reshuffled while the origin and end location of each trajectory are maintained.
}
\examples{
Randomize(species_df)
Randomize(species_df, RandTraj=500, grid=4, Plot=TRUE)
}
