% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomResults.R
\docType{data}
\name{randomResults}
\alias{randomResults}
\title{Example output from Randomise to demonstrate the PlotRandomTracks function}
\format{
List containing 4 list elements, based on \code{\link{Randomise}}
}
\usage{
randomResults
}
\description{
Output from \code{\link{Randomise}} function calculated using the \code{\link{tracks}} dataset and \code{\link{Randomise}} with randTrack=1. Note that the default is
normally randTrack=100; however, this creates an unnecessarily large file for an example dataset so randtrack=1 was used instead.
}
\keyword{datasets}
