% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\docType{data}
\name{tracks}
\alias{tracks}
\title{Sample location data to demonstrate PhysMove functions}
\format{
A data frame with 15623 rows and 4 variables:
\describe{
  \item{ref}{trajectory's unique id number as an integer}
  \item{lon}{longitude of each position estimate in decimal degrees}
  \item{lat}{latitide of each position estimate in decimal degrees}
  \item{day}{datetime stamp for each location estimate in POSIXct format following: yyyy-mm-dd hh:mm:ss}
  ...
}
}
\source{
\url{https://github.com/HannahCalich/PhysMove/tree/master/data-raw}
}
\usage{
tracks
}
\description{
A data frame containing location data from 25 sample trajectories.
}
\keyword{datasets}
