% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracksCRW.R
\docType{data}
\name{tracksCRW}
\alias{tracksCRW}
\title{Sample location data to demonstrate PhysMove functions}
\format{
A data frame with 15623 rows and 4 variables:
\describe{
  \item{ref}{trajectory ID number as an integer}
  \item{lon}{longitude of each position estimate in decimal degrees}
  \item{lat}{latitude of each position estimate in decimal degrees}
  \item{day}{datetime stamp for each location estimate in POSIXct format}
}
}
\usage{
tracksCRW
}
\description{
A data frame containing location data from 25 sample trajectories that were developed
to follow a correlated random walk model using the aniMotum R package.
}
\details{
Reference: https://ianjonsen.r-universe.dev/aniMotum
}
\keyword{datasets}
