% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angleList.R
\docType{data}
\name{angleList}
\alias{angleList}
\title{Example output from TurningAngles to demonstrate the PlotAngles function}
\format{
List containing 10 list elements, 1 for each time window set using \code{\link{TurningAngles}}
}
\usage{
angleList
}
\description{
Output from \code{\link{TurningAngles}} function calculated using the tracks dataset and \code{\link{TurningAngles}} with max_hr = 24. Note that the TurningAngles
default is normally max_hr=240; however, this creates an unnecessarily large file for an example dataset so max_hr = 24 was used instead.
}
\keyword{datasets}
