% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angleList.R
\docType{data}
\name{angleList}
\alias{angleList}
\title{Example output from TurningAngles to demonstrate the PlotAngles function}
\format{
List containing 10 list elements, 1 for each time window set using \code{\link{TurningAngles}} default settings
}
\usage{
angleList
}
\description{
Output from \code{\link{TurningAngles}} function calculated using the tracks dataset and default \code{\link{TurningAngles}} parameters
}
\keyword{datasets}
