% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTracks.R
\name{checkTracks}
\alias{checkTracks}
\title{Check data format
This function checks the format of telemetry data prior to running PhysMove metrics.}
\usage{
checkTracks(species_df)
}
\arguments{
\item{species_df}{A data frame containing telemetry data with columns named
ref, lon, lat, and day.}
}
\value{
Messages indicating whether the data are correctly formatted; warnings describe any issues.
}
\description{
Check data format
This function checks the format of telemetry data prior to running PhysMove metrics.
}
\details{
The columns must be formatted as follows:
ref: numeric ID for each individual.
lon and lat: numeric longitude and latitude in decimal degrees.
day: POSIXct datetime values.
Datetime format: %Y-%m-%d %H:%M:%S
}
\examples{
checkTracks(tracks)
}
