% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareDistributions_20210629.R
\name{CompDist}
\alias{CompDist}
\title{Identify the best-fit distribution for displacement data using weighted Akaike information criterion}
\usage{
CompDist(Displacements, force_AICc = FALSE)
}
\arguments{
\item{Displacements}{Input the output from the \code{\link{CalculateDisplacements}} function. This function also requires outputs from the \code{\link{FitDist}} function}

\item{Force_AICc}{Force function to calculate AICc scores instead of AIC scores when n/K is > 40. Default is FALSE.}
}
\value{
Updated "DistResults" data frame with summary statistics, AICc (or AIC) scores, and weighted AIC scores for each distribution fit.
}
\description{
This function allows you to identify the distribution that best-fits the pdf of the displacements, best-fit x-min, and parameters calculated using the \code{\link{CalculateDisplacements}} and \code{\link{FitDist}} functions. This function will automatically calculate AICc scores if n/K is <= 40 for the largest value of K, where n = sample size (N_Tail) and K = number of parameters in the model. See Burnham and Anderson (2004) for further details (DOI: 10.1177/0049124104268644).
}
\examples{
CompDist(Displacements)
}
