% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareDistributions_20210319.R
\name{CompDist}
\alias{CompDist}
\title{Identify the best-fit distribution for displacement data using weighted Akaike information criterion}
\usage{
CompDist(Displacements)
}
\arguments{
\item{Displacements}{Input the output from \code{\link{CalculateDisplacements}} function. To fit distribution to all displacements input MyDisp as is, else you can subset for temporal periods (e.g., MyDisp[1]).}
}
\value{
"DistResults_AIC" Dataframe with summary statistics and AIC scores for each distribution fit.
}
\description{
This function allows you to identify the distribution that best-fits the pdf of the normalized displacements calculated using \code{\link{CalculateDisplacements}}.
}
\examples{
CompDist(Displacements)
}
