% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompDist.R
\name{CompDist}
\alias{CompDist}
\title{Identify the best-fit distribution for displacement data}
\usage{
CompDist(displacements, distResults, force_AICc = FALSE)
}
\arguments{
\item{displacements}{List of displacements output from the \code{\link{CalcDisp}} function.}

\item{distResults}{Data frame of results output from the \code{\link{FitDist}} function.}

\item{force_AICc}{Force function to calculate AICc scores instead of AIC scores when n/K is > 40. Default is FALSE.}
}
\value{
A data frame with that contains the summary statistics for each distribution fit (from the \code{\link{FitDist}} function) as well as
the AICc/AIC scores and weighted AICc/AIC scores (wAICc/wAIC) for each distribution fit.
}
\description{
This function allows you to determine if a power law, exponential, or log-normal distribution best-fits a probability density function
of the displacements using weighted Akaike Information Criterion (AIC). These fits use displacement data, xmin and parameter values that were
previously calculated with the \code{\link{CalcDisp}} and \code{\link{FitDist}} functions. By default, this function will calculate
AICc scores (AIC scores corrected for small sample sizes) if n/K is <= 40 for the largest value of K, where n = sample size (nTail) and K = number of parameters in the model
(see Burnham and Anderson (2004) for further details, DOI: 10.1177/0049124104268644). However, if force_AICc = TRUE AICc scores will be
calculated regardless of n/K.
}
\examples{
CompDist(displacements)
CompDist(displacements, force_AICc=FALSE)
}
