% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disp.R
\docType{data}
\name{disp}
\alias{disp}
\title{Example output from calcDisp()}
\format{
list
}
\usage{
disp
}
\description{
Example output from the \code{\link{calcDisp}} function. Displacements calculated using
the \code{\link{tracks}} dataset and calcDisp() with max_hr=24. Note that the
calcDisp() default is normally max_hr=240; however, this creates an unnecessarily
large file for an example dataset so max_hr=24 was used instead.
}
\keyword{datasets}
