% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disp.R
\docType{data}
\name{disp}
\alias{disp}
\title{Example output from CalcDisp()}
\format{
A list with 10 elements
}
\usage{
disp
}
\description{
Example output from the \code{\link{CalcDisp}} function. Displacements in km calculated using the
\code{\link{tracks}} dataset and CalcDisp() with max_hr=24. Note that the CalcDisp() default is
normally max_hr=240; however, this creates an unnecessarily large file for an example dataset
so max_hr=24 was used instead.
}
\keyword{datasets}
