% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disp.R
\docType{data}
\name{disp}
\alias{disp}
\title{Sample displacement data to demonstrate PhysMove functions}
\format{
A list with 10 elements
}
\usage{
disp
}
\description{
A list containing the displacements in km recorded for each time window using CalcDisp default parameters
}
\keyword{datasets}
