% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distResultsAll.R
\docType{data}
\name{distResultsAll}
\alias{distResultsAll}
\title{Example output from FitDist function}
\format{
list
}
\usage{
distResultsAll
}
\description{
Output from \code{\link{FitDist}} function using tracks dataset and \code{\link{FitDist}} with full=TRUE. First list element contains a dataframe of results output from \code{\link{FitDist}}
and the second list element is a record of if the data were normalized or not, which is needed for \code{\link{CompDist}} and \code{\link{PlotDist}}).
}
\keyword{datasets}
