% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distResultsAll.R
\docType{data}
\name{distResultsAll}
\alias{distResultsAll}
\title{Example output from fitDist()}
\format{
list
}
\usage{
distResultsAll
}
\description{
Example output from the \code{\link{fitDist}} function. Full distribution fits calculated using the \code{\link{tracks}}
dataset and fitDist() with full=TRUE. The first list element contains a data frame of distribution results (dmin, parameters, etc)
and the second list element is a record of if the data were normalized or not, which is needed for \code{\link{compDist}}
and \code{\link{plotDist}}.
}
\keyword{datasets}
