% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distResultsExp.R
\docType{data}
\name{distResultsExp}
\alias{distResultsExp}
\title{Example output from FitDist function with set_dmin=1.649160 (best-fit d min for exponential distribution)}
\format{
list
}
\usage{
distResultsExp
}
\description{
Output from \code{\link{FitDist}} function using \code{\link{tracks}} dataset. All \code{\link{FitDist}} function parameters were default except for
set_dmin=1.649160, which is the best-fit d min for an exponential distribution using \code{\link{tracks}} dataset and default displacement calculations.
First list element contains a dataframe of results output from \code{\link{FitDist}} and the second list element is a record of if the data were normalized
or not, which is needed for \code{\link{CompDist}} and \code{\link{PlotDist}}).
}
\keyword{datasets}
