% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distResultsExp.R
\docType{data}
\name{distResultsExp}
\alias{distResultsExp}
\title{Example output from fitDist() using the best-fit dmin for an exponential distribution}
\format{
list
}
\usage{
distResultsExp
}
\description{
Example output from the \code{\link{fitDist}} function. Distribution fits calculated using the \code{\link{tracks}} dataset
and fitDist() with set_dmin=1.649160, which is the best-fit dmin for an exponential distribution for the \code{\link{tracks}} dataset.
The first list element contains a data frame of distribution results (dmin, parameters, etc) and the second list element is a record
of if the data were normalized or not, which is needed for \code{\link{compDist}} and \code{\link{plotDist}}.
}
\keyword{datasets}
