% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entropy_20210318.R
\name{Entropy}
\alias{Entropy}
\title{Entropy}
\usage{
Entropy(
  species_df,
  gridcell = 0.25,
  hist = TRUE,
  pdfplot = FALSE,
  nb = 40,
  legend = c(TRUE, "topleft")
)
}
\arguments{
\item{species_df}{A data frame containing location data (rows) and columns with the following headers: "ref", "lon", "lat", "day". "ref" is the unique id number for each animal
(e.g., their satellite tag number; integer format). "lon" and "lat" are the longitude and latitide of each position estimate in decimal degrees (numeric format). "day"
is the datetime stamp for each location estimate (POSIXct format following yyyy-mm-dd hh:mm:ss). See XXX data frame for an example.}

\item{nb}{Number of bins, used to determine the size of the bins used to calculate entropy frequency for plots. Default is 40.}

\item{grid}{Grid used to split 1-degree by 1-degree cells. A grid of 4 results in 0.25-degree by 0.25 degree grid cells, while a grid of 0.25 results in 4-degrees by 4-degree grid cells. Default is 4.}

\item{plot}{Create line plot of each individual's Entropy. Default is TRUE.}
}
\value{
Ind_Entropy vector containing Individual entropy values, NormalizedEntropy vector containing normalized values per individual, line plot (if desired).
}
\description{
This function allows you to calculate the entropy of individual trajectories, which provides insights to how ordered (or disordered) a trajectory was.
}
\examples{
Entropy(species_df)
Entropy(species_df,grid=4,plot=TRUE,nb=40)
}
