% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entropy.R
\name{Entropy}
\alias{Entropy}
\title{Entropy of trajectories}
\usage{
Entropy(
  species_df,
  gridCell = 0.25,
  histPlot = TRUE,
  legend = c(TRUE, "topleft"),
  pdfPlot = FALSE,
  nBins = 40
)
}
\arguments{
\item{species_df}{A data frame containing location data in rows. Columns have the following headers: "ref", "lon", "lat", "day".
"ref" is the unique id number for each animal (e.g., their satellite tag number formatted as an integer),
"lon" and "lat" are the longitude and latitude of each position estimate in decimal degrees in numeric format),
"day" is the datetime stamp for each location estimate in POSIXct format following yyyy-mm-dd hh:mm:ss.
See attached sample data \code{\link{plSample}}, \code{\link{expSample}}, or \code{\link{lnormSample}}.}

\item{gridCell}{Grid cell size in degrees. Default is 0.25.}

\item{histPlot}{Plot a histogram of the normalized  entropy values. Default is TRUE.}

\item{legend}{Add a legend to the histPlot when TRUE and change the position of the legend. Default is legend=c(TRUE, "topleft").}

\item{pdfPlot}{Create a  probability density function line plot of the entropy values. Default is TRUE.}

\item{nBins}{Number of bins used to calculate the pdf plot. Default is 40.}
}
\value{
Data vector ('normalizedEntropy') containing normalized entropy values for each individual, a histogram plot of the normalized entropy values
(if histPlot=TRUE), a  probability density function line plot and a data frame with the data used to create the pdf plot ('entropyPDFplot', if pdfPlot=TRUE),
and data needed for the Predictability function (a list of occurrences and a vector of raw entropy values, 'occurrences' and 'indivEntropy', respectively).
}
\description{
This function allows you to calculate the entropy of individual trajectories, which provides insight to how ordered (or disordered) the trajectories were.
}
\examples{
Entropy(expSample)
Entropy(expSample, gridCell=0.25, histPlot=TRUE, legend=c(TRUE, "topleft"), pdfPlot=FALSE, nBins=40)
}
