% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropyResults.R
\docType{data}
\name{entropyResults}
\alias{entropyResults}
\title{Example output from entropy()}
\format{
data.frame
}
\usage{
entropyResults
}
\description{
Example output from the \code{\link{entropy}} function. Results calculated using the
\code{\link{tracks}} dataset with entropy() default parameters
}
\keyword{datasets}
