% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FittingDistributions_20210319.R
\name{FitDist}
\alias{FitDist}
\title{Fitting Distributions to CDFs of Displacements}
\usage{
FitDist(
  Displacements,
  dist = c("pl", "exp", "lnorm"),
  set_xmin = NULL,
  Full = FALSE,
  Normalize = TRUE
)
}
\arguments{
\item{Displacements}{Input the output from \code{\link{CalculateDisplacements}} function. To fit distribution to all displacements input Displacements as is, else you can subset for temporal periods (e.g., Displacements[1]).}

\item{dist}{Continuous distributions to be fit. Possible values are power-law ("pl"), exponential ("exp), or lognormal ("lnorm") continuous distributions. Default is all three distributions (i.e., dist=c("pl","exp","lnorm")).}

\item{set_xmin}{To limit the fitted distribution to values above a specified xmin. Default set_xmin=NULL.}

\item{Full}{To fit the fitted distribution to the full range of displacement data. Default Full=FALSE.Fit}

\item{Normalize}{Normalizes the displacement distances by dividing each displacement by the average displacement for that time period. Required if working with displacements calculated over multiple time windows.}

\item{AIC}{Calculate the weighted AIC values of the distribution fits to identify the best-fit distribution for your data. Note that AIC values are only meaningful if you are comparing model fits over the same distribution of data. Default AIC=FALSE}

\item{Plot}{Plot the Complementary Cumulative Distribution Function for the displacements and add fit lines for each calculated distribution, if desired. Default is c(TRUE, TRUE).}
}
\value{
Dataframe with summary statistics for each distribution fit.
}
\description{
This function allows you to identify the distribution that best-fits the pdf of the normalized displacements calculated using \code{\link{CalculateDisplacements}}.To begin, this function
normalizes all of the displacements by dividing each displacement by the mean displacement from it's corresponding time period and combines the results into a vector of Normalized values.
}
\examples{
CalculateDisplacements(species_df, dist=c("pl","exp"))
CalculateDisplacements(species_df, dist=c("pl","exp","lnorm"), set_xmin=NULL, Full=FALSE, AIC=TRUE, Normalize=TRUE)
}
