% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDist.R
\name{fitDist}
\alias{fitDist}
\title{Fit distributions to data}
\usage{
fitDist(
  input,
  dist = c("pl", "exp", "lnorm"),
  set_xmin = NULL,
  full = FALSE,
  normalise = TRUE
)
}
\arguments{
\item{input}{List of values used to fit distribution}

\item{dist}{Continuous distributions that will be fit to the data. Possible values are power law ("pl"), exponential ("exp"), or log-normal ("lnorm")
continuous distributions. Default is dist=c("pl","exp","lnorm").}

\item{set_xmin}{To limit the fitted distribution to values above a specified value. If your data are going to be normalised
this value will have to be a normalised value as well. Default is NULL.}

\item{full}{To fit the distributions to the full range of data. Default is FALSE.}

\item{normalise}{Normalises the displacement distances by dividing each displacement by the average displacement for that time window;
normalise=TRUE is required if working with data calculated over multiple time windows.}
}
\value{
A list including a dataframe of summary statistics for each distribution fit (1st list element). Results dataframe includes the
distribution name, xmin (minimum value used to fit each distribution), parameter 1 (alpha, lambda, mu) and parameter 2 (NA, NA, sigma) for pl, exp, and lnorm
distributions respectively, and nTail (the number of data points greater than or equal to xmin). A logical argument indicating if
data were normalised is exported as the 2nd list element because this information is needed for the \code{\link{compDist}} and \code{\link{plotDist}}
functions.
}
\description{
This function allows you to fit power law, exponential, or lognormal distributions to data. For example, this function can be used
to fit data output from the \code{\link{calcDisp}} function or the \code{\link{occupancy}} function.
}
\examples{
fitDist(input, dist=c("pl","exp","lnorm"), full=TRUE)
}
