% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\name{FitDist}
\alias{FitDist}
\title{Fit distributions to displacements}
\usage{
FitDist(
  displacements,
  dist = c("pl", "exp", "lnorm"),
  set_dmin = NULL,
  full = FALSE,
  normalise = TRUE
)
}
\arguments{
\item{displacements}{List of displacements output from the \code{\link{CalcDisp}} function.}

\item{dist}{Continuous distributions that will be fit to the displacements. Possible values are power law ("pl"), exponential ("exp"), or log-normal ("lnorm")
continuous distributions. Default is dist=c("pl","exp","lnorm").}

\item{set_dmin}{To limit the fitted distribution to values above a specified value. If your displacements are going to be normalised
this value will have to be a normalised value as well. Default is NULL.}

\item{full}{To fit the distributions to the full range of displacement data. Default is FALSE.}

\item{normalise}{Normalises the displacement distances by dividing each displacement by the average displacement for that time window;
normalise=TRUE is required if working with displacements calculated over multiple time windows.}
}
\value{
A list including a dataframe of summary statistics for each distribution fit (1st list element). Results dataframe includes the
distribution name, dmin (d values used to fit each distribution), parameter 1 (alpha, lambda, mu) and parameter 2 (NA, NA, sigma) for pl, exp, and lnorm
distributions respectively, and nTail (the number of data points greater than or equal to dmin). A logical argument indicating if
data were normalised is exported as the 2nd list element because this information is needed for the \code{\link{CompDist}} and \code{\link{PlotDist}}
functions.
}
\description{
This function allows you to fit power law, exponential, or lognormal distributions to the displacements calculated with
the \code{\link{CalcDisp}} function.
}
\examples{
FitDist(disp, dist=c("pl","exp","lnorm"), full=TRUE)
}
