% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rG_20210315.R
\name{GyrationRad}
\alias{GyrationRad}
\title{Gyration Radius}
\usage{
GyrationRad(
  species_df,
  map = TRUE,
  Colours = c("Black", "Red"),
  plot = FALSE,
  nb = 15
)
}
\arguments{
\item{species_df}{A data frame containing location data (rows) and columns with the following headers: "ref", "lon", "lat", "day". "ref" is the unique id number for each animal
(e.g., their satellite tag number; integer format). "lon" and "lat" are the longitude and latitide of each position estimate in decimal degrees (numeric format). "day"
is the datetime stamp for each location estimate (POSIXct format following yyyy-mm-dd hh:mm:ss). See XXX data frame for an example.}

\item{map}{Create a map illustrating gyration radii. Default is TRUE.}

\item{Colours}{Colours for points and gyration radius, respectively. Default is c("Black","Red).}

\item{plot}{Line plot illustrating the probability density function of gyration radii scores.  Default is TRUE.}

\item{nb}{Number of bins, used to determine the size of the bins used to calculate predictability frequencies for plots. Default is 40.}
}
\value{
Vector containing gyration radius values for each trajectory, line plot, and/or map (if desired).
}
\description{
This function allows you to calculate the gyration radius of individual trajectories.
}
\examples{
rG(species_df)
rG(species_df,plot=TRUE,nb, map=TRUE)
}
