% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gyrationRad.R
\name{gyrationRad}
\alias{gyrationRad}
\title{Gyration Radius}
\usage{
gyrationRad(species_df, map = TRUE, mapCol = c("Black", "Red"))
}
\arguments{
\item{species_df}{A data frame containing location data in rows. Columns have the following headers: "ref", "lon", "lat", "day".
"ref" is the unique id number for each animal (e.g., their satellite tag number formatted as an integer),
"lon" and "lat" are the longitude and latitude of each position estimate in decimal degrees in numeric format,
"day" is the datetime stamp for each location estimate in POSIXct format following '%Y-%m-%d %H:%M:%S'.
See attached sample data \code{\link{tracks}}.}

\item{map}{Create a map illustrating the gyration radius of each trajectory. Default is TRUE.}

\item{mapCol}{Colours for points and gyration radii on map, respectively. Default is c("Black","Red").}
}
\value{
A data frame containing the unique trajectory identifier (ref), the mean location (longitude and latitude), and
the gyration radius (rG, in km) for each trajectory. If map = TRUE, a map of the gyration radius results is also produced
}
\description{
This function calculates the gyration radius of individual trajectories.
A pdf plot of the gyration radius values can be created with the \code{\link{plotPDF}} function.
}
\examples{
\dontrun{

gyrationRad(tracks, map=TRUE, mapCol=c("Black","Red"))
}
}
