% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infomapCommunities.R
\name{infomapCommunities}
\alias{infomapCommunities}
\title{Identify Infomap communities and create a transition probability matrix}
\usage{
infomapCommunities(
  species_df,
  gridCell = 0.25,
  hours = 24,
  range_hr = 6,
  tpm = FALSE
)
}
\arguments{
\item{species_df}{A data frame containing location data in rows. Columns have the following headers: "ref", "lon", "lat", "day".
"ref" is the unique id number for each animal (e.g., their satellite tag number formatted as an integer),
"lon" and "lat" are the longitude and latitude of each position estimate in decimal degrees in numeric format,
"day" is the datetime stamp for each location estimate in POSIXct format following '%Y-%m-%d %H:%M:%S'.
See attached sample data \code{\link{tracks}}.}

\item{gridCell}{Grid cell size in degrees. Default is 0.25.}

\item{hours}{Identify locations separated by this number of hours for movement calculations. Default is 24.}

\item{range_hr}{Range (in hours) converts the hours parameter into a time window (hours +/-  range_hr) so the
code can identify location estimates that are close to, but not exactly separated by a set number of hours.
If multiple location estimates fall within this time window the location estimate closest to the set hours input value
will be used for calculations. For example, if hours = 24 and range = 6, the algorithm will search for
locations spaced 18 to 30 hours apart. Default is 6.}

\item{tpm}{Export the transition probability matrix in link list format. If tpm=TRUE, a 'TransitionProbabilityMatrix' data frame will be automatically
returned as the second element of the output list. Default is FALSE.}
}
\value{
A list where element 1 ('infomap_object') contains the Infomap results summarising the hierarchical structure of communities.
If tpm = TRUE, element 2 ('tpm') contains the transition probability matrix used to construct the network.
The transition probability matrix is returned in link list format (origin node, destination node, and transition probability).
}
\description{
This function uses the network community detection Infomap to identify Infomap communities based on a transition probability matrix (tpm), which summarizes
the probability of individuals moving from one grid cell to another over a set time window. This function assumes directed movement, allows for self-links
(where an individual stays in the same cell over time), and uses a tpm in link list format to create an Infomap 'monolayer_object'. Note that if warnings
appear about columns or rows summing to 0 this simply means an individual moved into a cell and did not leave, which is a valid movement and not cause for alarm.
}
\details{
Please note: to run this function you must first download the infomapecology and emln R packages from GitHub and install the stand-alone Infomap file.
For details please see: https://ecological-complexity-lab.github.io/infomap_ecology_package/installation
To learn more about Infomap please visit: https://www.mapequation.org/

Example:
library (infomapecology)
infomapCommunities(tracks, gridCell=0.25, hours=24, range_hr=6, tpm=FALSE)
}
