% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infomapResult.R
\docType{data}
\name{infomapResult}
\alias{infomapResult}
\title{Example output from InfomapCommunities to demonstrate the CommunityMap function}
\format{
Infomap monolayer object
}
\usage{
infomapResult
}
\description{
Infomap monolayer object created using the InfomapCommunities function with default parameters and the tracks dataset
}
\keyword{datasets}
