% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Occupancy.R
\name{Occupancy}
\alias{Occupancy}
\title{Occupancy}
\usage{
Occupancy(
  species_df,
  gridCell = 0.25,
  map = TRUE,
  colGrad = c("blue", "light blue", "red")
)
}
\arguments{
\item{species_df}{A data frame containing location data in rows. Columns have the following headers: "ref", "lon", "lat", "day".
"ref" is the unique id number for each animal (e.g., their satellite tag number formatted as an integer),
"lon" and "lat" are the longitude and latitude of each position estimate in decimal degrees in numeric format),
"day" is the datetime stamp for each location estimate in POSIXct format following yyyy-mm-dd hh:mm:ss.
See attached sample data \code{\link{tracks}}.}

\item{gridCell}{Grid cell size in degrees. Default is 0.25.}

\item{map}{Create a map illustrating where occupancy occurs. Default is TRUE.}

\item{colGrad}{Colour gradient for occupancy map that illustrates low, moderate, and high occupancy, respectively
(applied to ggplot2::scale_fill_gradientn). Default is colGrad=c("blue", "light blue","red").}
}
\value{
A dataframe including occupancy values and corresponding
locations (provided as centres value of each grid cell). If map=TRUE a map is created.
}
\description{
This function allows you to calculate the occupancy patterns of location estimates and create a map.
A pdf plot of the occupancy values can be created with the \code{\link{PlotPDF}} function.
}
\examples{
Occupancy(tracks)
Occupancy(tracks, gridCell=0.25, map=TRUE, colGrad=c("blue", "light blue", "red"))
}
