% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupancyResults.R
\docType{data}
\name{occupancyResults}
\alias{occupancyResults}
\title{Example output from occupancy()}
\format{
dataframe
}
\usage{
occupancyResults
}
\description{
Example output from the \code{\link{occupancy}} function. Results calculated using the \code{\link{tracks}} dataset
and occupancy() default parameters.
}
\keyword{datasets}
