% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDist.R
\name{PlotDist}
\alias{PlotDist}
\title{Plot best-fit distributions to complementary cumulative distribution function (CCDF) of displacements}
\usage{
PlotDist(
  displacements,
  fitLines = TRUE,
  colours = c("red", "gold2", "blue"),
  setDist = NULL,
  legend = c(TRUE, "bottomleft")
)
}
\arguments{
\item{displacements}{The output from the \code{\link{CalcDisp}} function.}

\item{fitLines}{Add fit lines based on the parameters calculated with the \code{\link{FitDist}} function. Default is TRUE.}

\item{colours}{Colours for each fit line. Default is colours=c("red","gold2","blue").}

\item{setDist}{Plot  a subset of the distribution fit lines calculated with the \code{\link{FitDist}} function. To plot a specific
distribution fit line use setDist="pl", or to plot multiple distributions use setDist=c("pl","exp"). Options include "pl", "exp", and "lnorm".
Default is NULL}

\item{legend}{Add a legend to the plot when TRUE and change the position of the legend. Default is legend=c(TRUE, "bottomleft")}
}
\value{
Complementary cumulative distribution function (CCDF) plot of displacements with fit lines (if fitLines=TRUE).
}
\description{
This function allows you to plot a complementary cumulative distribution function (CCDF) of displacements with fit lines
based on the distribution fits calculated with the \code{\link{FitDist}} function.
}
\examples{
PlotDist(displacements, fitLines=TRUE, colours=c("red","gold2","blue"), setDist=FALSE, legend=c(TRUE, "bottomleft"))
PlotDist(displacements)
}
