% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDist.R
\name{plotDist}
\alias{plotDist}
\title{Plot best-fit distributions to complementary cumulative distribution function (ccdf) of data}
\usage{
plotDist(
  input,
  distResults,
  fitLines = TRUE,
  setDist = NULL,
  colours = c("red", "gold2", "blue"),
  legend = TRUE,
  label = NULL
)
}
\arguments{
\item{input}{List of values used to fit distributions.}

\item{distResults}{List output from the \code{\link{fitDist}} function containing a dataframe of fit results (list element 1) and a normalisation record (list element 2)}

\item{fitLines}{Add fit lines based on the parameters calculated with the \code{\link{fitDist}} function. Default is TRUE.}

\item{setDist}{Plot a subset of lines for each distribution fit calculated with the \code{\link{fitDist}} function (e.g., setDist=c("pl","exp"))
Options include "pl", "exp", and "lnorm". The lines will be drawn in order from "pl", then "exp", then "lnorm" (when applicable).
By default all lines are plotted. Default is NULL.}

\item{colours}{Colours for each fit line. Valid input options include colour names or hex numbers. Default is colours=c("red","gold2","blue").}

\item{legend}{Add legend with legend=TRUE. Default is TRUE.}

\item{label}{X axis label. Note that "Normalised" will automatically be added if distributions were fit to normalised data. Default is NULL and will result in x-axis label of "input data".}
}
\value{
Complementary cumulative distribution function (ccdf) plot of input with fit lines (if fitLines=TRUE).
}
\description{
This function allows you to plot a complementary cumulative distribution function (ccdf) of values with fit lines
based on distribution fits calculated with the \code{\link{fitDist}} function.
}
\examples{
plotDist(disp, distResultsExp)
}
