% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDist.R
\name{PlotDist}
\alias{PlotDist}
\title{Plot best-fit distributions to complementary cumulative distribution function (ccdf) of displacements}
\usage{
PlotDist(
  displacements,
  distResults,
  fitLines = TRUE,
  setDist = NULL,
  colours = c("red", "gold2", "blue"),
  legend = TRUE
)
}
\arguments{
\item{displacements}{List of displacements output from the \code{\link{CalcDisp}} function.}

\item{distResults}{List output from the \code{\link{FitDist}} function containing a dataframe of fit results (list element 1) and a normalisation record (list element 2)}

\item{fitLines}{Add fit lines based on the parameters calculated with the \code{\link{FitDist}} function. Default is TRUE.}

\item{setDist}{Plot a subset of lines for each distribution fit calculated with the \code{\link{FitDist}} function (e.g., setDist=c("pl","exp"))
Options include "pl", "exp", and "lnorm". The lines will be drawn in order from "pl", then "exp", then "lnrom" (when applicable).
By default all lines are plotted. Default is NULL.}

\item{colours}{Colours for each fit line. The colours correspond to the drawing order: "pl", "exp", "lnorm" (when applicable).
Valid input options include colour names or hex numbers. Default is colours=c("red","gold2","blue").}

\item{legend}{Add legend with legend=TRUE. Default is TRUE.}
}
\value{
Complementary cumulative distribution function (ccdf) plot of displacements with fit lines (if fitLines=TRUE).
}
\description{
This function allows you to plot a complementary cumulative distribution function (ccdf) of displacements with fit lines
based on the displacements output from the \code{\link{CalcDisp}} function and the distribution fits calculated with the
\code{\link{FitDist}} function.
}
\examples{
PlotDist(displacements, distResultsExp)
}
