% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPDF.R
\name{PlotPDF}
\alias{PlotPDF}
\title{Plot a probability density function}
\usage{
PlotPDF(result, desc = NULL, nBins)
}
\arguments{
\item{result}{Data used to create plot.}

\item{desc}{Description of input data. This parameter is used to determine how the data are plotted and to assign appropriate x and y plot labels.
Valid input options include: "Occupancy" (e.g., from the \code{\link{Occupancy}} function), "GyrationRad"  (e.g., from the \code{\link{GyrationRad}} function),
"Entropy" (e.g., from the \code{\link{Entropy}} function), and "Predictability" (from the \code{\link{Predictability}} function). Occupancy pdfs are based on
the total number of cells globally based on the provided grid cell size and are created on created on a log-log scale due to the nature of the data ranges while the other desc types are create on a standard xy plot (to plot occupancy on a standard
xy scale leave desc as default). Default is NULL.}

\item{nBins}{Number of bins used to calculate the pdf plot (e.g., nBins=25). By default, if desc="Occupancy" the code will use 20 log-sized bins (due to the
nature of the data ranges) else the number of bins is determined by the data ranges. If the input result values range from 0 to 1 (e.g., entropy or
predictability results) the code will automatically use 40 bins, and if the input results fall outside the 0 to 1 range (e.g., gyration radius results) the
code will automatically use 15 bins.}
}
\value{
A pdf plot of the results and the data used to create the plot.
}
\description{
This function allows you to plot a probability density function (pdf).
}
\examples{
PlotPDF(occupancyResultsDF$Occupancy, desc="Occupancy")
}
