% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reshuffling_PlotTracks_20210317.R
\name{PlotRandomTracks}
\alias{PlotRandomTracks}
\title{Randomized Trajectories}
\usage{
PlotRandomTracks(
  species_df,
  Ref,
  RandTraj = 10,
  colours = c("black", "red"),
  pchtype = c(19, 19),
  title = "",
  legend = c(TRUE, "topleft")
)
}
\arguments{
\item{species_df}{A data frame containing location data (rows) and columns with the following headers: "ref", "lon", "lat", "day". "ref" is the unique id number for each animal
(e.g., their satellite tag number; integer format). "lon" and "lat" are the longitude and latitide of each position estimate in decimal degrees (numeric format). "day"
is the datetime stamp for each location estimate (POSIXct format following yyyy-mm-dd hh:mm:ss). See XXX data frame for an example.}

\item{Ref}{Trajectory ID number from species_df to plot}

\item{RandTraj}{Random trajectory(ies) to plot. Values can include either 1 number, or a range of numbers following #:# format. Default is 1:10.}

\item{Colours}{Colours for real trajectory locations (Colour option 1) and reshuffled trajectory locations (colour option 2). Default is Colours=c("black","red").}

\item{Legend}{Add legend to plot and specify location. Legend=c(TRUE, "topleft").}
}
\value{
Plot showing real and reshuffled trajectory locations.
}
\description{
Plot locations from real and reshuffled trajectories using RandomizedLat and RandomizedLong outputs from \code{\link{Randomize}} function
}
\examples{
PlotReshuffTracks<-function(species_df, Ref=33933, RandomizedLat, RandomizedLong, RandTraj=1)
PlotReshuffTracks(species_df, Ref=33933, RandomizedLat, RandomizedLong, RandTraj=1:10, Colours=c("black","red"))
}
