% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRandomTracks.R
\name{PlotRandomTracks}
\alias{PlotRandomTracks}
\title{Plot Randomised Tracks}
\usage{
PlotRandomTracks(
  species_df,
  ref = NULL,
  numPlot = 1:5,
  colours = c("black", "grey70"),
  tracks = TRUE,
  startCol = "red",
  endCol = "blue",
  legend = TRUE
)
}
\arguments{
\item{species_df}{A data frame containing location data in rows. Columns have the following headers: "ref", "lon", "lat", "day".
"ref" is the unique id number for each animal (e.g., their satellite tag number formatted as an integer),
"lon" and "lat" are the longitude and latitude of each position estimate in decimal degrees in numeric format),
"day" is the datetime stamp for each location estimate in POSIXct format following yyyy-mm-dd hh:mm:ss.
See attached sample data \code{\link{tracks}}.}

\item{ref}{Reference number of track from species_df to plot.}

\item{numPlot}{Number of Randomised tracks to plot. The Randomised tracks were consecutively numbered from 1 to however many you set in the
\code{\link{Randomise}} function. The input value can either be any of these individual numbers (e.g., 23), or a range of numbers (e.g., 1:10),
which will plot all of the random tracks created within the range. Default is 1:5.}

\item{colours}{Colours to plot points from original and Randomised tracks, respectively. Default is colours=c("black","grey70").}

\item{tracks}{Add track lines to the plot. Default is TRUE.}

\item{startCol}{Colour for origin location. startCol=NULL will cause the symbology of the origin location to match the symbology of the rest
of the original track. Default is startCol="red".}

\item{endCol}{Colour for destination location. endCol=NULL will cause the symbology of the destination location to match the symbology of the rest
of the original track. Default is endCol="blue".}

\item{legend}{legend Add legend with legend=TRUE (default).}
}
\value{
Plot showing the original and Randomised track locations and the Randomised tracks data used to create the map (original tracks are
from species_df).
}
\description{
Plot locations from original and reshuffled tracks using RandomisedLat and RandomisedLong outputs from the \code{\link{Randomise}} function
}
\examples{
PlotRandomTracks<-function(tracks, ref=1)
PlotRandomTracks<-(tracks, ref=1, numPlot=1:5, colours=c("black","grey70"), tracks=TRUE, startCol="red", endCol="blue", legend=TRUE)
}
