% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictability.R
\name{predictability}
\alias{predictability}
\title{Predictability of trajectories}
\usage{
predictability(species_df, entropyResults, startVal = NULL, histPlot = TRUE)
}
\arguments{
\item{species_df}{A data frame containing location data in rows. Columns have the following headers: "ref", "lon", "lat", "day".
"ref" is the unique id number for each animal (e.g., their satellite tag number formatted as an integer),
"lon" and "lat" are the longitude and latitude of each position estimate in decimal degrees in numeric format,
"day" is the datetime stamp for each location estimate in POSIXct format following '%Y-%m-%d %H:%M:%S'.
See attached sample data \code{\link{tracks}}.}

\item{entropyResults}{Data frame of results output from the \code{\link{entropy}} function.}

\item{startVal}{Optional starting value used to find a root value for the limit of predictability equation.
If NULL (default), the starting value is automatically determined from the normalised entropy for each individual.
The function will iteratively decrease the starting value by 0.01 until an acceptable root within (0,1) is found.}

\item{histPlot}{Plot a histogram of the limit of predictability scores. Default is TRUE.}
}
\value{
Limit of predictability values for each trajectory. If histPlot=TRUE a histogram of the limit of predictability scores is created.
}
\description{
This function allows you to calculate the limit of predictability for each trajectory based on each individual's entropy. This function requires
'indivEntropy', 'cellsVisited', and 'normalisedEntropy' from the \code{\link{entropy}} function.
A pdf plot of the predictability values can be created with the \code{\link{plotPDF}} function.
}
\examples{
predictability(tracks, entropyResults, startVal = NULL, histPlot=TRUE)
}
