% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Randomise.R
\name{Randomise}
\alias{Randomise}
\title{Randomise tracks}
\usage{
Randomise(species_df, randTrack = 500, gridCell = 0.25, plot = TRUE, lm = TRUE)
}
\arguments{
\item{species_df}{A data frame containing location data in rows. Columns have the following headers: "ref", "lon", "lat", "day".
"ref" is the unique id number for each animal (e.g., their satellite tag number formatted as an integer),
"lon" and "lat" are the longitude and latitude of each position estimate in decimal degrees in numeric format),
"day" is the datetime stamp for each location estimate in POSIXct format following yyyy-mm-dd hh:mm:ss.
See attached sample data \code{\link{tracks}}.}

\item{randTrack}{Number of randomised tracks per individual. Default is 500.}

\item{gridCell}{Grid cell size in degrees. Default is 0.25.}

\item{plot}{Plot the number of cells visited in the original track versus the average number of cells visited in the
reshuffled tracks. Default is TRUE.}

\item{lm}{Calculate a linear regression to examine the relationship between the number of cells visited in the original tracks
(target variable) and the average number of cells visited by the Randomised tracks (predictor variable). If plot = TRUE this parameter
adds a solid black fit line to the data points and a black dashed line, which represents a 1:1 relationship. The slope of the fit line can
be determined by typing 'RandomiselinearModel$coefficients[2]'. Default is TRUE.}
}
\value{
The number of cells visited by each original track and the average number of cells visited by the Randomised tracks, for each
ref. If plot = TRUE, a plot illustrating the number of cells visited by the original and Randomised tracks is created.
If lm = TRUE, a linear regression is run, a fit line and reference line are added to the plot (if plot = TRUE), and the results
'RandomiselinearModel' are automatically assigned to the global environment. Three additional variables are automatically assigned
to the global environment as they are required to plot the reshuffled tracks with the \code{\link{PlotRandomTracks} function}
('RandomisedLat' and 'RandomisedLong').
}
\description{
This function allows you to investigate the influence spatial and/or temporal correlations may have on an indiivudals' space use patterns.
This is done by maintaining the origin and end location of each track, randomizing the order displacements occurred in between these points,
and calculating how many grid cells the original and Randomised tracks visited, which summarizes their space use.
}
\examples{
Randomise(tracks)
Randomise(tracks, randTrack=500, gridCell=0.25, plot=TRUE, lm=TRUE)
}
