% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Randomise.R
\name{Randomise}
\alias{Randomise}
\title{Randomise tracks}
\usage{
Randomise(species_df, randTrack = 100, gridCell = 0.25, plot = TRUE, lm = TRUE)
}
\arguments{
\item{species_df}{A data frame containing location data in rows. Columns have the following headers: "ref", "lon", "lat", "day".
"ref" is the unique id number for each animal (e.g., their satellite tag number formatted as an integer),
"lon" and "lat" are the longitude and latitude of each position estimate in decimal degrees in numeric format,
"day" is the datetime stamp for each location estimate in POSIXct format following yyyy-mm-dd hh:mm:ss.
See attached sample data \code{\link{tracks}}.}

\item{randTrack}{Number of randomised tracks per individual. Default is 100.}

\item{gridCell}{Grid cell size in degrees. Default is 0.25.}

\item{plot}{Plot the number of cells visited in the original track versus the average number of cells visited in the
reshuffled tracks. Default is TRUE.}

\item{lm}{Calculate a linear regression to examine the relationship between the number of cells visited in the original tracks
(target variable) and the average number of cells visited by the Randomised tracks (predictor variable). If plot = TRUE this parameter
adds a solid black fit line to the data points and a black dashed line, which represents a 1:1 relationship. The slope of the fit line can
be determined by typing 'RandomiselinearModel$coefficients[2]'. Default is TRUE.}
}
\value{
List containing a dataframe of results (list element 1), the randomised longitude and latitude values (list elements 2 and 3, respectively),
which are needed for the \code{\link{PlotRandomTracks}} function, and if lm = TRUE, the results of the linear model are output (list element 4).
The dataframe of results includes columns for the number of cells visited by each original track and the average number of cells visited by the
Randomised tracks for each ref. Lastly, if plot = TRUE, a plot illustrating the number of cells visited by the original and randomised tracks is created,
and if lm = TRUE, a fit line and reference line are added to the plot.
}
\description{
This function allows you to investigate the influence spatial and/or temporal correlations may have on an individuals' space use patterns.
This is done by maintaining the origin and end location of each track, randomizing the order displacements occurred in between these points,
and calculating how many grid cells the original and randomised tracks visited, which summarizes their space use.
}
\examples{
\dontrun{

Randomise(tracks, randTrack=100, gridCell=0.25, plot=TRUE, lm=TRUE)

}
}
