% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomResults.R
\docType{data}
\name{randomResults}
\alias{randomResults}
\title{Example output from Randomise to demonstrate the PlotRandomTracks function}
\format{
List containing 4 list elements, based on \code{\link{Randomise}} default parameters
}
\usage{
randomResults
}
\description{
Output from \code{\link{Randomise}} function calculated using the \code{\link{tracks}} dataset and default \code{\link{Randomise}} parameters
}
\keyword{datasets}
