% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomResults.R
\docType{data}
\name{randomResults}
\alias{randomResults}
\title{Example output from randomise()}
\format{
list
}
\usage{
randomResults
}
\description{
Example output from the \code{\link{randomise}} function. Results calculated using the \code{\link{tracks}} dataset
and randomise() with randTrack=1. Note that the randomise() default is normally randTrack=100; however, this creates
an unnecessarily large file for an example dataset so randTrack=1 was used instead.
}
\keyword{datasets}
