% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMS_20210316.R
\name{RMS}
\alias{RMS}
\title{Root-Mean-Square of Displacement}
\usage{
RMS(
  species_df,
  timeunit = "hours",
  bw = 1.1,
  plot = TRUE,
  pchtype = c(15, 19),
  colours = c("black", "red"),
  legend = c(TRUE, "topleft"),
  lm = FALSE
)
}
\arguments{
\item{species_df}{A data frame containing location data (rows) and columns with the following headers: "ref", "lon", "lat", "day". "ref" is the unique id number for each animal
(e.g., their satellite tag number; integer format). "lon" and "lat" are the longitude and latitide of each position estimate in decimal degrees (numeric format). "day"
is the datetime stamp for each location estimate (POSIXct format following yyyy-mm-dd hh:mm:ss). See XXX data frame for an example.}

\item{timeunit}{Unit used to calculate time between locations (e.g., "days", "hours", "seconds"). Default is "hours".}

\item{bw}{Bin width refers to the size of the bins used to calculate how frequently displacements occurred. Altering this value will change the number of points on your plot. Default is 1.5.}

\item{plot}{Plot the root-mean-square of displacements and the mean displacements against their corresponding time periods. Default is TRUE.}

\item{pchtype}{Point type for plot. Default is 19}

\item{colours}{Colours for plot points. Default is c("black", "red").}

\item{legend}{Adds legend to plot and specifies legend location. Default is c(TRUE, "bottomleft").}

\item{lm}{Run a linear model}
}
\value{

}
\description{
This function allows you to calculate the root-mean-square of displacements and plot them scaled with time
}
\examples{
RMS(species_df)
RMS(species_df,bw=1.5,plot=TRUE,colours=c("black","red"),legend=c(TRUE, "topleft"))
}
