% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Angles_20210317.R
\name{TurningAngles}
\alias{TurningAngles}
\title{Calculate turning angles from trajectories}
\usage{
TurningAngles(
  species_df,
  min_hr = 24,
  max_hr = 240,
  interval_hr = 24,
  range_hr = 6,
  spiderPlot = c(TRUE, "all"),
  histPlot = c(FALSE, "all"),
  colours = "Navy"
)
}
\arguments{
\item{species_df}{A data frame containing location data in rows. Columns have the following headers: "ref", "lon", "lat", "day".
"ref" is the unique id number for each animal (e.g., their satellite tag number formatted as an integer),
"lon" and "lat" are the longitude and latitude of each position estimate in decimal degrees in numeric format),
"day" is the datetime stamp for each location estimate in POSIXct format following yyyy-mm-dd hh:mm:ss.
See attached sample data \code{\link{plSample}}, \code{\link{expSample}}, or \code{\link{lnormSample}}.}

\item{min_hr}{Minimum number of hours to consider for calculations. Default is 24 hours.}

\item{max_hr}{Maximum number of hours to consider for calculations. Default is 240 hours.}

\item{interval_hr}{Time interval (in hours) used to set intervals between min_hr and max_hr. Default is 24 hours.}

\item{range_hr}{Range (in hours) converts interval_hr into a time window (interval_hr +/-  range_hr) so the
code can identify location estimates that are close to, but not exactly separated by the interval_hr input value.
If multiple location estimates fall within this time window the location estimate closest to the interval_hr input value
will be used for calculations. For example, if interval_hr = 24 and range = 6, the algorithm will search for
locations spaced 18 to 32 hours apart. Default is 6.}

\item{spiderPlot}{Plot a spider plot showing the frequency of turning angles from either all time windows combined
(default) or one specific time period. For example, spiderPlot=c(TRUE,1) to plot only the first time period. Default spiderPlot=c(TRUE, "all").}

\item{histPlot}{Plot a histogram showing the frequency of turning angles from all time windowscombined (default) or
one specific time period. For example, histPlot=c(TRUE,1) to plot only the first time period. Default is histPlot=c(TRUE, "all").}

\item{colours}{Colour(s) for bars in histPlot and lines in spiderPlot. Default is "Navy".}
}
\value{
List of Turning Angles for each time window where each list element corresponds with the time windows set.
If histPlot and/or spiderPlot=TRUE, a histogram and/or spiderPlot of results are exported.
}
\description{
This function allows you to calculate turning angles between sets of three consecutive location estimates separated by set time period(s).
}
\examples{
TurningAngles(expSample)
TurningAngles(expSample, min_hr=24, max_hr=240, interval_hr=24,range_hr=6, spiderPlot=c(TRUE, "all"), histPlot=c(FALSE, "all"), colours="Navy")
}
