% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Angles_20210317.R
\name{TurningAngles}
\alias{TurningAngles}
\title{Calculating Turning Angles from Trajectories}
\usage{
TurningAngles(
  species_df,
  min_hr = 24,
  max_hr = 240,
  interval_hr = 24,
  range_hr = 6,
  spiderplot = c(TRUE, "all"),
  histplot = c(FALSE, "all"),
  colours = "Navy"
)
}
\arguments{
\item{species_df}{A data frame containing location data (rows) and columns with the following headers: "ref", "lon", "lat", "day". "ref" is the unique id number for each animal
(e.g., their satellite tag number; integer format). "lon" and "lat" are the longitude and latitide of each position estimate in decimal degrees (numeric format). "day"
is the datetime stamp for each location estimate (POSIXct format following yyyy-mm-dd hh:mm:ss). See XXX data frame for an example.}

\item{min_hr}{Minimum number of hours to consider for displacement calculations. Default is 24 hours.}

\item{max_hr}{Maximum number of hours to consider for displacement calculations (default is 240 hours)}

\item{interval_hr}{Time interval (in hours) used to identify time period intervals between min_hr and max_hr (default is 24 hours)}

\item{range_hr}{Range (in hours) applied to interval_hr. This value helps the algorithm identify location estimates that are close to, but not exactly separated by the interval_hr. If multiple location estimates fall within this range the location estimate closest to the interval_hr will be used for calculations.}

\item{spiderplot}{Plot a spider plot showing the frequency of turning angles from either all time periods combined or one specific time period. For example, spiderplot=c(TRUE,1) to plot only the 1st time period. Default spiderplot=c(TRUE, "all").}

\item{histplot}{Plot a histogram showing the frequency of turning angles from either all time periods combined or one specific time period. For example, histplot=c(TRUE,1) to plot only the 1st time period. Default histplot=c(TRUE, "all").}

\item{colours}{Colour(s) for bars in histplot and lines in spiderplot. Default is "Navy".}
}
\value{
List of Turning Angles for each time period and a histogram and/or spiderplot of results, if desired
}
\description{
To calculate turning angles between sets of three consecutive location estimates separated by set time period(s).
}
\examples{
TurningAngles(species_df, max_hr=24)
TurningAngles(species_df,min_hr=24,max_hr=240,interval_hr=24,range_hr=6, plot=TRUE)
}
